+++ include/spdk/bdev_module.h
@@ -17,6 +17,7 @@
 
 #include "spdk/bdev.h"
 #include "spdk/bdev_zone.h"
+#include "spdk/log.h"
 #include "spdk/queue.h"
 #include "spdk/scsi_spec.h"
 #include "spdk/thread.h"